+++ src/main/groovy/com/ullink/NuGetPush.groovy
@@ -1,25 +1,53 @@
 package com.ullink
 
+
+import org.gradle.api.tasks.Input
+import org.gradle.api.tasks.InputFile
+import org.gradle.api.tasks.Optional
+
 class NuGetPush extends BaseNuGet {
 
-    def nupkgFile
-
+    @Optional
+    @InputFile
+    File nupkgFile
+    @Optional
+    @Input
     def serverUrl
+    @Optional
+    @InputFile
     def apiKey
-    def timeout
+    @Optional
+    @InputFile
     def configFile
 
     NuGetPush() {
         super('push')
+
+        // Force always execute
+        outputs.upToDateWhen { false }
     }
 
-    String getNugetPackOutputFile() {
+    void setNupkgFile(String path) {
+        nupkgFile = project.file(path)
+    }
+
+    void setApiKey(String path) {
+        apiKey = project.file(path)
+    }
+
+    void setConfigFile(String path) {
+        configFile = project.file(path)
+    }
+
+    @Optional
+    @InputFile
+    File getNugetPackOutputFile() {
         if (dependentNuGetPack)
            dependentNuGetPack.packageFile
     }
 
     NuGetPack getDependentNuGetPack() {
-        dependsOn.find { it instanceof NuGetPack }
+        dependsOn.find { it instanceof NuGetPack } as NuGetPack
     }
 
     @Override
@@ -28,7 +56,6 @@ class NuGetPush extends BaseNuGet {
 
         if (serverUrl) args '-Source', serverUrl
         if (apiKey) args '-ApiKey', apiKey
-        if (timeout) args '-Timeout', timeout
         if (configFile) args '-ConfigFile', configFile
 
         super.exec()