+++ src/test/groovy/com/ullink/MSBuildTaskBuilder.groovy
@@ -1,53 +1,51 @@
 package com.ullink
 
-import org.gradle.api.Task
+import org.gradle.api.Project
 
 class MSBuildTaskBuilder {
-
-    private final Task msbuildTask
     private final Map<String, Object> mainProjectProperties = [:]
+    private final Project project
     private final mainProject = new Object()
     private final List<File> artifacts = []
 
-    public MSBuildTaskBuilder() {
-        msbuildTask = [
-                getName: { 'msbuild' }
-        ] as Task
-        mainProject.metaClass.getProperties = { mainProjectProperties }
-        mainProject.metaClass.getDotnetArtifacts = { artifacts }
-        msbuildTask.metaClass.getMainProject = { mainProject }
-        msbuildTask.metaClass.parseProject = true
+    MSBuildTaskBuilder(Project project) {
+        this.project = project
     }
 
-    public Task build() {
-        msbuildTask
+    def build() {
+        mainProject.metaClass.dotnetArtifacts = artifacts
+        mainProject.metaClass.properties = mainProjectProperties
+        project.tasks.register('msbuild') {
+            it.metaClass.mainProject = mainProject
+            it.metaClass.parseProject = true
+        }
     }
 
-    public MSBuildTaskBuilder withAssemblyName(String assemblyName) {
+    MSBuildTaskBuilder withAssemblyName(String assemblyName) {
         mainProjectProperties['AssemblyName'] = assemblyName
         this
     }
 
-    public MSBuildTaskBuilder withFrameworkVersion(String version) {
+    MSBuildTaskBuilder withFrameworkVersion(String version) {
         mainProjectProperties['TargetFrameworkVersion'] = version
         this
     }
 
-    public MSBuildTaskBuilder withArtifact(String artifactPath) {
+    MSBuildTaskBuilder withArtifact(String artifactPath) {
         artifacts.add(new File(artifactPath))
         this
     }
 
-    public MSBuildTaskBuilder withProjectFile(String path) {
+    MSBuildTaskBuilder withProjectFile(String path) {
         mainProject.metaClass.getProjectFile = { path }
         this
     }
 
-    public MSBuildTaskBuilder withProjectFile(File file) {
+    MSBuildTaskBuilder withProjectFile(File file) {
         withProjectFile(file.path)
     }
 
-    public MSBuildTaskBuilder withMainProjectProperty(String name, String value) {
+    MSBuildTaskBuilder withMainProjectProperty(String name, String value) {
         mainProjectProperties[name] = value
         this
     }