+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # gradle-nuget-plugin changelog
 
+## 2.19
+### Changed
+* Add Input and Output Annotations for Gradle 7 [#77](https://github.com/Itiviti/gradle-nuget-plugin/issues/77)
+* Built with gradle 6.2.2
+* Default NuGet version used is 5.5.0
+
+### Removed
+* Due to conflict in Gradle 6.1, `timeout` in `nugetPush` is removed [#78](https://github.com/Itiviti/gradle-nuget-plugin/issues/78)
+* Usage of deprecated base.getDistsDir [#76](https://github.com/Itiviti/gradle-nuget-plugin/issues/76)
+
 ## 2.18
 ### Changed
 * built with gradle 4.7 (above won't work till we migrate out of plugindev)